+++ testdata/apps/cors/app/api/middleware.go
@@ -1,0 +1,5 @@
+package api
+
+import "github.com/emersonjoe/trilha"
+
+func MiddlewareOPTIONS(c *trilha.Ctx, next trilha.Next) error { return next() }