+++ testdata/apps/err_unroutable_method/app/arquivos/route.go
@@ -1,0 +1,7 @@
+package arquivos
+
+import "github.com/emersonjoe/trilha"
+
+func GET(c *trilha.Ctx) error { return c.Text(200, "x") }
+
+func HEAD(c *trilha.Ctx) error { return c.Text(200, "") }